mdz/pytorch/bert_cls/3_deploy/modelzoo/bert_cls/requirements.txt
torch==1.9.0
transformers==4.31.0
typing_extensions==4.8.0
numpy==1.24.3
PyYAML==6.0.1
boto3==1.35.45
regex==2.5.132